improves code colocation (list -> card)
