+++ src/component/List/index.js
@@ -2,22 +2,11 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { v4 as uuid } from 'uuid';
 import Card from '../Card';
-import {
-  removeItemById,
-  addCardToState,
-  deleteCardFromState,
-  updateCardFromState,
-  updateLists,
-} from '../../utils/index';
+import { removeItemById, addCardToState, updateLists } from '../../utils/index';
 
 export default function List({ list, setBoardState }) {
   const [addCardInput, setAddCardInput] = useState('');
   const updateCardInput = e => setAddCardInput(e.target.value);
-
-  // Drag and drop function
-  function onDragStart(e) {
-    e.dataTransfer.setData('draggedCardData', e.currentTarget.id);
-  }
 
   // Drag and drop function
   function onDragOver(e) {
@@ -28,21 +17,13 @@ export default function List({ list, setBoardState }) {
   function onDrop(e, targetList) {
     e.preventDefault();
     const cardInfo = e.dataTransfer.getData('draggedCardData');
-    setBoardState(state => updateLists(state, JSON.parse(cardInfo), targetList));
+    setBoardState(state => updateLists(state, cardInfo, targetList));
   }
 
   const onAddNewCard = e => {
     e.preventDefault();
     setBoardState(state => addCardToState(state, list.id, { title: addCardInput, id: uuid() }));
     setAddCardInput('');
-  };
-
-  const onDeleteCard = cardId => {
-    setBoardState(state => deleteCardFromState(state, list.id, cardId));
-  };
-
-  const onUpdateCard = (cardId, newTitle) => {
-    setBoardState(state => updateCardFromState(state, list.id, cardId, newTitle));
   };
 
   function deleteList(listId) {
@@ -53,14 +34,7 @@ export default function List({ list, setBoardState }) {
     <div className="list" onDragOver={onDragOver} onDrop={e => onDrop(e, list.id)}>
       <h3 className="list-title ">{list.listName}</h3>
       {list.cards.map(card => (
-        <Card
-          key={card.id}
-          card={card}
-          list={list}
-          onDragStart={onDragStart}
-          onDeleteCard={onDeleteCard}
-          onUpdateCard={onUpdateCard}
-        />
+        <Card key={card.id} card={card} list={list} setBoardState={setBoardState} />
       ))}
       <form className="list__add-card" onSubmit={onAddNewCard}>
         <input type="text" value={addCardInput} onChange={updateCardInput} />
